Separated public key tests into a test module
