+++ pallets/msa/src/tests/creation_tests.rs
@@ -38,3 +38,15 @@ fn it_throws_msa_identifier_overflow() {
 		assert_noop!(Msa::create(test_origin_signed(1)), Error::<Test>::MsaIdOverflow);
 	});
 }
+
+#[test]
+#[allow(unused_must_use)]
+fn it_does_not_allow_duplicate_keys() {
+	new_test_ext().execute_with(|| {
+		Msa::create(test_origin_signed(1));
+
+		assert_noop!(Msa::create(test_origin_signed(1)), Error::<Test>::KeyAlreadyRegistered);
+
+		assert_eq!(Msa::get_current_msa_identifier_maximum(), 1);
+	});
+}
