+++ pallets/msa/src/tests/mod.rs
@@ -4,6 +4,7 @@ mod creation_tests;
 mod delegation_tests;
 mod governance_tests;
 mod other_tests;
+mod public_key_tests;
 mod replay_tests;
 mod retirement_tests;
 mod signature_registry_tests;
